transactionID: int -> long

+++ src/lockmgr/DeadlockException.java
@@ -4,14 +4,14 @@ package lockmgr;
  * Thrown to indicate that the transaction is deadlocked and should be aborted.
  */
 public class DeadlockException extends Exception {
-    private int tid;
+    private long tid;
 
-    public DeadlockException(int tid, String msg) {
+    public DeadlockException(long tid, String msg) {
         super("The transaction " + tid + " is deadlocked:" + msg);
         this.tid = tid;
     }
 
-    int tid() {
+    long tid() {
         return tid;
     }
 }
